app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
click==8.1.7
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
pydantic==2.10.0
pydantic_core==2.27.0
python-dotenv==1.0.1
python-status==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
kafka-python==2.0.2
